% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_treasury_yields.R
\name{get_treasury_yields}
\alias{get_treasury_yields}
\title{Fetch Treasury Yield Data}
\usage{
get_treasury_yields()
}
\value{
A tibble containing datetime and numeric values
}
\description{
Returns data in either tidy or time-series format
}
